refactor(nametag): move to engine_graphics, remove Renderer* parameter
- Move nametag_system from engine_scene to engine_graphics to break
  circular dependency (scene_renderer -> nametag_system)
- Remove Renderer* parameter from DrawNametags, use ServiceLocator
  instead (consistent with Network/AssetManager pattern)
- Remove dead engine_dummy.cpp (never compiled, CH_ENGINE_SHARED=OFF)
