--- engine/engine_dummy.cpp
@@ -1,3 +1,0 @@
-void Engine_Dummy()
-{
-}
